fix:codex - review fixes

+++ cmd/addie/flags/stats.go
@@ -45,7 +45,7 @@ func statsFlags(expertMode bool) []cli.Flag {
 			Hidden:   expertMode,
 			Value:    5 * time.Second,
 			Action: validate("stats-metrics-interval", toSecondsCeil,
-				v.Required, v.Min(float64(5))),
+				v.Required, v.Min(5)),
 			// TODO : MINOR : add mod(v) == 0 check
 		},
 		&cli.DurationFlag{
